pyproject.toml
[tool.poetry]
name = "pan-os-upgrade"
version = "1.3.0"
description = "Python script to automate the upgrade process of PAN-OS firewalls."
authors = ["Calvin Remsburg <cremsburg.dev@gmail.com>"]
documentation = "https://cdot65.github.io/pan-os-upgrade/"
license = "Apache 2.0"
readme = "README.md"
packages = [{ include = "pan_os_upgrade" }]
keywords = ["paloaltonetworks", "panos", "firewall", "upgrade", "automation"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Framework :: Flake8",
  "Framework :: Pydantic :: 2",
  "Framework :: MkDocs",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: Apache Software License",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3.11",
  "Topic :: System :: Networking :: Firewalls",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: System :: Systems Administration",
  "Topic :: Utilities",
]
include = ["pan_os_upgrade/assets/*.png"]
exclude = [
  "pan_os_upgrade/assurance/*",
  "pan_os_upgrade/logs/*",
  "pan_os_upgrade/*.yaml",
]


[tool.poetry.dependencies]
python = "^3.11"
pan-os-python = "^1.11.0"
panos-upgrade-assurance = "^0.3.3"
pydantic = "^2.5.3"
typer = { extras = ["all"], version = "^0.9.0" }
setuptools = "^69.0.3"
dnspython = "^2.5.0"
dynaconf = "^3.2.4"
pyyaml = "^6.0.1"
reportlab = "^4.0.9"
prompt-toolkit = "^3.0.43"
tabulate = "^0.9.0"
colorama = "^0.4.6"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
flake8 = "^7.0.0"
ipdb = "^0.13.13"
mkdocs-material = "^9.5.4"
mkdocstrings = "^0.24.0"
pytest = "^7.4.4"
python-dotenv = "^1.0.1"
pytest-mock = "^3.12.0"
prompt-toolkit = "^3.0.43"

[tool.poetry.scripts]
pan-os-upgrade = 'pan_os_upgrade.main:app'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
